+++ src/DomainEvent.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace OG\Core\Domain;
+
+/**
+ * Event for notifying that something happened in the application domain.
+ */
+interface DomainEvent
+{
+}