+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Interfaces for use-case architecture
 - Domain exception
+- Interface for domain events
 
 ### Changed
 - Nothing